+++ Deep-EIoU/reid/torchreid/utils/rerank.py
@@ -40,11 +40,11 @@ def re_ranking(q_g_dist, q_q_dist, g_g_dist, k1=20, k2=6, lambda_value=0.3):
         ],
         axis=0
     )
-    original_dist = np.power(original_dist, 2).astype(float32)
+    original_dist = np.power(original_dist, 2).astype(float)
     original_dist = np.transpose(
         1. * original_dist / np.max(original_dist, axis=0)
     )
-    V = np.zeros_like(original_dist).astype(float32)
+    V = np.zeros_like(original_dist).astype(float)
     initial_rank = np.argsort(original_dist).astype(np.int32)
 
     query_num = q_g_dist.shape[0]
